Makefile: Avoid double macro expansion in targets
POSIX specification stays:

        string1 = [string2]
        ...
        Macro expansions in string1 of macro definition lines shall
        be evaluated when read. Macro expansions in string2 of macro
        definition lines shall be performed when the macro identified
        by string1 is expanded in a rule or command.

It means that recursive macro expansion is not guaranteed to work in
a portable Make. Also, as make is a declarative language makes more
sense to declare your targets as a primary concern instead of
derivating them from a informational macro like SRC that is only
used in a rule command.